fix(datetime): align epoch Go types with declared Sig() TypeInt
- toEpoch Eval: cast int64 -> int to match descriptor 'int' return type
- fromEpoch Eval: use coerce.ToInt() instead of coerce.ToInt64() to match
  Sig() TypeInt declaration; cast int -> int64 only for time.UnixMilli call
- Update tests: assert int (not int64) return from toEpoch; pass int to fromEpoch

+++ function/datetime/epoch.go
@@ -26,7 +26,7 @@ func (fnToEpoch) Sig() (paramTypes []data.Type, isVariadic bool) {
 	return []data.Type{data.TypeString}, false
 }
 
-// Eval converts a datetime string to Unix epoch milliseconds (int64).
+// Eval converts a datetime string to Unix epoch milliseconds.
 // datetime.toEpoch("2024-01-15T10:30:00Z") => 1705314600000
 func (fnToEpoch) Eval(params ...interface{}) (interface{}, error) {
 	log.RootLogger().Debugf("datetime.toEpoch: params=%+v", params)
@@ -34,7 +34,7 @@ func (fnToEpoch) Eval(params ...interface{}) (interface{}, error) {
 	if err != nil {
 		return nil, fmt.Errorf("datetime.toEpoch: argument is not a valid datetime: %v", params[0])
 	}
-	millis := t.UnixMilli()
+	millis := int(t.UnixMilli())
 	log.RootLogger().Debugf("datetime.toEpoch: parsed=%v epoch_ms=%d", t, millis)
 	return millis, nil
 }
@@ -54,11 +54,11 @@ func (fnFromEpoch) Sig() (paramTypes []data.Type, isVariadic bool) {
 // datetime.fromEpoch(1705314600000) => "2024-01-15T10:30:00Z"
 func (fnFromEpoch) Eval(params ...interface{}) (interface{}, error) {
 	log.RootLogger().Debugf("datetime.fromEpoch: params=%+v", params)
-	ms, err := coerce.ToInt64(params[0])
+	ms, err := coerce.ToInt(params[0])
 	if err != nil {
 		return nil, fmt.Errorf("datetime.fromEpoch: argument must be an integer (epoch milliseconds), got %v", params[0])
 	}
-	t := time.UnixMilli(ms).UTC()
+	t := time.UnixMilli(int64(ms)).UTC()
 	result := t.Format(time.RFC3339)
 	log.RootLogger().Debugf("datetime.fromEpoch: epoch_ms=%d result=%q", ms, result)
 	return result, nil
